refactor : concert 및 reservation 도메인 로직 리팩터링
