+++ src/main/java/kr/hhplus/be/server/api/concert/presentation/controller/SwaggerApi.java
@@ -212,5 +212,5 @@ public interface SwaggerApi {
                                     )
                             ))
             })
-    RestResponse<ConcertResponse.ReservedSeatInfo> payConcertSeat (ConcertRequest.ReserveConcert request);
+    RestResponse<ConcertResponse.ReservedSeatInfo> payConcertSeat (ConcertRequest.PayReserveConcert request);
 }
