+++ src/main/java/kr/hhplus/be/server/api/concert/presentation/dto/ConcertRequest.java
@@ -4,6 +4,13 @@ public class ConcertRequest {
     public record ReserveConcert(
             long userId,
             long scheduleId,
+            long seatId
+    ) {
+    }
+
+    public record PayReserveConcert(
+            long userId,
+            long scheduleId,
             long seatId,
             long reservedId
     ) {
